+++ app/(case-studies)/case-studies/components/data.ts
