+++ app/(main)/[lang]/data.ts
